Move article page progress bar to its own component
